consolidate getClient() and getUserID() in a Utils.ts file
