+++ src/ServiceAccountList.tsx
@@ -8,7 +8,7 @@ import { Link } from "react-router-dom"
 
 import { authn, meta } from "@brigadecore/brigade-sdk"
 
-import getClient from "./Client"
+import { getClient } from "./Utils"
 import LockIcon from "./components/LockIcon"
 import withPagingControl from "./components/PagingControl"
